+++ dplanner/src/main/java/com/dp/dplanner/service/CommentService.java
@@ -6,9 +6,9 @@ import com.dp.dplanner.domain.CommentMemberLike;
 import com.dp.dplanner.domain.Post;
 import com.dp.dplanner.domain.club.ClubMember;
 import com.dp.dplanner.dto.CommentMemberLikeDto;
-import com.dp.dplanner.exception.clubMember.ClubMemberException;
-import com.dp.dplanner.exception.comment.CommentException;
-import com.dp.dplanner.exception.post.PostException;
+import com.dp.dplanner.exception.ClubMemberException;
+import com.dp.dplanner.exception.CommentException;
+import com.dp.dplanner.exception.PostException;
 import com.dp.dplanner.repository.ClubMemberRepository;
 import com.dp.dplanner.repository.CommentMemberLikeRepository;
 import com.dp.dplanner.repository.CommentRepository;
