+++ dplanner/src/main/java/com/dp/dplanner/service/LockService.java
@@ -5,9 +5,9 @@ import com.dp.dplanner.domain.Lock;
 import com.dp.dplanner.domain.Period;
 import com.dp.dplanner.domain.Resource;
 import com.dp.dplanner.domain.club.ClubMember;
-import com.dp.dplanner.exception.clubMember.ClubMemberException;
-import com.dp.dplanner.exception.lock.LockException;
-import com.dp.dplanner.exception.resource.ResourceException;
+import com.dp.dplanner.exception.ClubMemberException;
+import com.dp.dplanner.exception.LockException;
+import com.dp.dplanner.exception.ResourceException;
 import com.dp.dplanner.repository.ClubMemberRepository;
 import com.dp.dplanner.repository.LockRepository;
 import com.dp.dplanner.repository.ResourceRepository;
