+++ dplanner/src/main/java/com/dp/dplanner/service/PostService.java
@@ -7,9 +7,9 @@ import com.dp.dplanner.domain.club.Club;
 import com.dp.dplanner.domain.club.ClubMember;
 import com.dp.dplanner.dto.AttachmentDto;
 import com.dp.dplanner.dto.PostMemberLikeDto;
-import com.dp.dplanner.exception.club.ClubException;
-import com.dp.dplanner.exception.clubMember.ClubMemberException;
-import com.dp.dplanner.exception.post.PostException;
+import com.dp.dplanner.exception.ClubException;
+import com.dp.dplanner.exception.ClubMemberException;
+import com.dp.dplanner.exception.PostException;
 import com.dp.dplanner.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
