+++ dplanner/src/main/java/com/dp/dplanner/service/ResourceService.java
@@ -4,9 +4,9 @@ import com.dp.dplanner.aop.annotation.RequiredAuthority;
 import com.dp.dplanner.domain.Resource;
 import com.dp.dplanner.domain.club.Club;
 import com.dp.dplanner.domain.club.ClubMember;
-import com.dp.dplanner.exception.club.ClubException;
-import com.dp.dplanner.exception.clubMember.ClubMemberException;
-import com.dp.dplanner.exception.resource.ResourceException;
+import com.dp.dplanner.exception.ClubException;
+import com.dp.dplanner.exception.ClubMemberException;
+import com.dp.dplanner.exception.ResourceException;
 import com.dp.dplanner.repository.ClubMemberRepository;
 import com.dp.dplanner.repository.ClubRepository;
 import com.dp.dplanner.repository.ResourceRepository;
