+++ dplanner/src/test/java/com/dp/dplanner/service/LockServiceTest.java
@@ -7,8 +7,8 @@ import com.dp.dplanner.domain.Resource;
 import com.dp.dplanner.domain.club.Club;
 import com.dp.dplanner.domain.club.ClubMember;
 import com.dp.dplanner.exception.BaseException;
-import com.dp.dplanner.exception.clubMember.ClubMemberException;
-import com.dp.dplanner.exception.lock.LockException;
+import com.dp.dplanner.exception.ClubMemberException;
+import com.dp.dplanner.exception.LockException;
 import com.dp.dplanner.repository.ClubMemberRepository;
 import com.dp.dplanner.repository.LockRepository;
 import com.dp.dplanner.repository.ResourceRepository;
