+++ dplanner/src/test/java/com/dp/dplanner/service/PostServiceTest.java
@@ -11,8 +11,8 @@ import com.dp.dplanner.dto.AttachmentDto;
 import com.dp.dplanner.dto.PostMemberLikeDto;
 import com.dp.dplanner.dto.Status;
 import com.dp.dplanner.exception.BaseException;
-import com.dp.dplanner.exception.clubMember.ClubMemberException;
-import com.dp.dplanner.exception.post.PostException;
+import com.dp.dplanner.exception.ClubMemberException;
+import com.dp.dplanner.exception.PostException;
 import com.dp.dplanner.repository.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
